refactor: update imports to use shadcn directory in storage locations pages
